+++ supertokens_python/recipe/session/utils.py
@@ -13,7 +13,11 @@
 # under the License.
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Callable, Union, Awaitable, Literal
+from typing import TYPE_CHECKING, Callable, Union, Awaitable
+try:
+    from typing import Literal
+except ImportError:
+    from typing_extensions import Literal
 from urllib.parse import urlparse
 
 from tldextract import extract